+++ report_template/src/App.js
@@ -1,7 +1,6 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
 
-import SideNav from "./topComponents/SideNav";
 import LoginMain from "./loginPage/Login";
 import FEARequest from "./requestPage/FeaRequest";
 import BaseQueue from "./basicQueuePage/BasicQueue";
@@ -10,7 +9,6 @@ import WorkQueue from "./workQueuePage/WorkQueue";
 function App() {
   return (
     <Router>
-      <SideNav />
       <Switch>
         <Route path="/" exact component={LoginMain} />
         <Route path="/request" component={FEARequest} />
