+++ report_template/src/comps_common/Header.js
@@ -2,7 +2,7 @@ import React from "react";
 
 const Header = ({ titleHeader }) => {
   return (
-    <h1 className=" text-4xl p-5 mb-5 bg-blue-900 w-full text-white font-bold">
+    <h1 className=" text-4xl p-5  bg-blue-900 w-full text-white font-bold">
       {titleHeader}
     </h1>
   );
