+++ report_template/src/workQueuePage/WorkQueue.js
@@ -1,10 +1,33 @@
 import React from "react";
 
+import Header from "../comps_common/Header";
+import SideNav from "../SideNavComp/SideNav";
+
+const QueuedItem = () => {
+  return (
+    <div className="bg-white flex shadow rounded  my-2">
+      <ul className="flex flex-row w-full justify-between px-2 py-5">
+        <li> Status of model </li>
+        <div>
+          <li> Model Name submitted </li>
+          <li> quick description of model </li>
+        </div>
+        <li> date:</li>
+        <li> button to update</li>
+      </ul>
+    </div>
+  );
+};
+
 export default function WorkQueue() {
   return (
-    <h1 text-4xl p-5 mb-5 bg-blue-900 w-full text-white font-bold>
-      {" "}
-      Here is the main working queue!!
-    </h1>
+    <div>
+      <Header titleHeader="V2 Request Queue" />
+      <SideNav />
+      <div className="flex flex-col w-full">
+        <QueuedItem />
+        <QueuedItem />
+      </div>
+    </div>
   );
 }
